refactor: split todo.command.ts (2924 lines) into focused modules
Closes #898
